requirements.txt
flask
flask_sqlalchemy
pymysql
flask_cors
python-dotenv